requirements.txt

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
playwright>=1.40.0
pillow>=10.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# LLM providers
anthropic>=0.7.0
openai>=1.0.0
groq>=0.4.0

# LangGraph and state management
langgraph>=0.0.40
langchain-core>=0.1.0
langchain>=0.1.0

# Computer vision and OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Utilities
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
